mpcontribs-portal/requirements.txt
beautifulsoup4==4.8.2
Django==3.0.5
django-extensions==2.2.9
django-settings-file==2.1.0
django-webpack4-loader==0.0.5
gunicorn[gevent]==20.0.4
ipykernel==5.2.0
jupyter-client==6.1.2
monty==3.0.2
nbconvert==5.6.1
nbformat==5.0.5
pandas-datareader==0.8.1
pymongo==3.10.1
scipy==1.4.1
whitenoise==5.0.1
